Rename src/lib/api to src/lib/hooks
